docker changes

+++ si_project_2025_be/database/migrations/2025_11_02_070017_create_contact_persons_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('phone')->nullable();
 
 
-            $table->integer('company_id')->index();
+            $table->unsignedBigInteger('company_id')->index();
 
             $table->foreign('company_id')
                 ->references('company_id')
